requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-dotenv==1.0.0
google-generativeai==0.3.1
aiohttp==3.9.1
redis==5.0.1
pytest==7.4.3
pydantic==2.5.0
